+++ main.go
@@ -429,9 +429,13 @@ func loadCustomMiddleware(referenceSpec *APISpec) ([]string, tykcommon.Middlewar
 	// Set AuthCheck hook
 	if referenceSpec.APIDefinition.CustomMiddleware.AuthCheck.Name != "" {
 		mwAuthCheckFunc = referenceSpec.APIDefinition.CustomMiddleware.AuthCheck
+		if referenceSpec.APIDefinition.CustomMiddleware.AuthCheck.Path != "" {
+			// Feed a JS file to Otto
+			mwPaths = append(mwPaths, referenceSpec.APIDefinition.CustomMiddleware.AuthCheck.Path)
+		}
 	}
 
-	// Load form the configuration
+	// Load from the configuration
 	for _, mwObj := range referenceSpec.APIDefinition.CustomMiddleware.Pre {
 		mwPaths = append(mwPaths, mwObj.Path)
 		mwPreFuncs = append(mwPreFuncs, mwObj)
@@ -477,6 +481,61 @@ func loadCustomMiddleware(referenceSpec *APISpec) ([]string, tykcommon.Middlewar
 		}
 	}
 
+
+	// Get Auth folder path
+	middlewareAuthFolderPath := path.Join(config.MiddlewarePath, referenceSpec.APIDefinition.APIID, "auth")
+	mwAuthFiles, _ := ioutil.ReadDir(middlewareAuthFolderPath)
+	for _, f := range mwAuthFiles {
+		if strings.Contains(f.Name(), ".js") {
+			filePath := filepath.Join(middlewareAuthFolderPath, f.Name())
+			log.WithFields(logrus.Fields{
+				"prefix": "main",
+			}).Debug("Loading Auth file middleware from ", filePath)
+			middlewareObjectName := strings.Split(f.Name(), ".")[0]
+			log.WithFields(logrus.Fields{
+				"prefix": "main",
+			}).Debug("-- Middleware name ", middlewareObjectName)
+
+			thisMWDef := tykcommon.MiddlewareDefinition{}
+			thisMWDef.Name = middlewareObjectName
+			thisMWDef.Path = filePath
+			thisMWDef.RequireSession = false
+
+			mwPaths = append(mwPaths, filePath)
+			mwAuthCheckFunc = thisMWDef
+			// only one allowed!
+			break
+		}
+	}
+
+	// Get POSTKeyAuth folder path
+	middlewarePostKeyAuthFolderPath := path.Join(config.MiddlewarePath, referenceSpec.APIDefinition.APIID, "post_auth")
+	mwPostKeyAuthFiles, _ := ioutil.ReadDir(middlewarePostKeyAuthFolderPath)
+	for _, f := range mwPostKeyAuthFiles {
+		if strings.Contains(f.Name(), ".js") {
+			filePath := filepath.Join(middlewarePostKeyAuthFolderPath, f.Name())
+			log.WithFields(logrus.Fields{
+				"prefix": "main",
+			}).Debug("Loading POST-KEY-AUTH-PROCESSOR file middleware from ", filePath)
+			middlewareObjectName := strings.Split(f.Name(), ".")[0]
+			log.WithFields(logrus.Fields{
+				"prefix": "main",
+			}).Debug("-- Middleware name ", middlewareObjectName)
+
+			requiresSession := strings.Contains(middlewareObjectName, "_with_session")
+			log.WithFields(logrus.Fields{
+				"prefix": "main",
+			}).Debug("-- Middleware requires session: ", requiresSession)
+			thisMWDef := tykcommon.MiddlewareDefinition{}
+			thisMWDef.Name = middlewareObjectName
+			thisMWDef.Path = filePath
+			thisMWDef.RequireSession = requiresSession
+
+			mwPaths = append(mwPaths, filePath)
+			mwPostKeyAuthFuncs = append(mwPostFuncs, thisMWDef)
+		}
+	}
+
 	// Get POST folder path
 	middlewarePostFolderPath := path.Join(config.MiddlewarePath, referenceSpec.APIDefinition.APIID, "post")
 	mwPostFiles, _ := ioutil.ReadDir(middlewarePostFolderPath)
@@ -512,6 +571,10 @@ func loadCustomMiddleware(referenceSpec *APISpec) ([]string, tykcommon.Middlewar
 
 	// Load PostAuthCheck hooks
 	for _, mwObj := range referenceSpec.APIDefinition.CustomMiddleware.PostKeyAuth {
+		if mwObj.Path != "" {
+			// Otto files are specified here
+			mwPaths = append(mwPaths, mwObj.Path)	
+		}
 		mwPostKeyAuthFuncs = append(mwPostKeyAuthFuncs, mwObj)
 	}
 
@@ -974,6 +1037,12 @@ func loadApps(APISpecs *[]*APISpec, Muxer *mux.Router) {
 
 				useCoProcessAuth := EnableCoProcess && mwDriver != tykcommon.OttoDriver && referenceSpec.EnableCoProcessAuth
 
+				var useOttoAuth bool = false
+				if !useCoProcessAuth {
+					useOttoAuth = mwDriver == tykcommon.OttoDriver && referenceSpec.EnableCoProcessAuth	
+				}
+				
+
 				if referenceSpec.APIDefinition.UseBasicAuth {
 					// Basic Auth
 					log.WithFields(logrus.Fields{
@@ -1012,7 +1081,7 @@ func loadApps(APISpecs *[]*APISpec, Muxer *mux.Router) {
 					// TODO: check if mwAuthCheckFunc is available/valid
 					log.WithFields(logrus.Fields{
 						"prefix": "main",
-					}).Info("----> Checking security policy: CoProcess")
+					}).Info("----> Checking security policy: CoProcess Plugin")
 
 					log.WithFields(logrus.Fields{
 						"prefix": "coprocess",
@@ -1024,7 +1093,15 @@ func loadApps(APISpecs *[]*APISpec, Muxer *mux.Router) {
 					}
 				}
 
-				if referenceSpec.UseStandardAuth || (!referenceSpec.UseOpenID && !referenceSpec.EnableJWT && !referenceSpec.EnableSignatureChecking && !referenceSpec.APIDefinition.UseBasicAuth && !referenceSpec.APIDefinition.UseOauth2 && !useCoProcessAuth) {
+				if useOttoAuth {
+					log.WithFields(logrus.Fields{
+						"prefix": "main",
+					}).Info("----> Checking security policy: JS Plugin")
+
+					authArray = append(authArray, CreateDynamicAuthMiddleware(mwAuthCheckFunc.Name, tykMiddleware))
+				}
+
+				if referenceSpec.UseStandardAuth || (!referenceSpec.UseOpenID && !referenceSpec.EnableJWT && !referenceSpec.EnableSignatureChecking && !referenceSpec.APIDefinition.UseBasicAuth && !referenceSpec.APIDefinition.UseOauth2 && !useCoProcessAuth && !useOttoAuth) {
 					// Auth key
 					log.WithFields(logrus.Fields{
 						"prefix": "main",