+++ rpc_storage_handler.go
@@ -48,6 +48,8 @@ var RPC_EmergencyModeLoaded bool
 
 var ErrorDenied error = errors.New("Access Denied")
 
+var GlobalRPCCallTimeout time.Duration
+
 // ------------------- CLOUD STORAGE MANAGER -------------------------------
 
 var RPCClients = map[string]chan int{}
@@ -231,7 +233,7 @@ func (r *RPCStorageHandler) GroupLogin() {
 		UserKey: r.UserKey,
 		GroupID: config.SlaveOptions.GroupID,
 	}
-	ok, err := r.Client.Call("LoginWithGroup", groupLoginData)
+	ok, err := r.Client.CallTimeout("LoginWithGroup", groupLoginData, GlobalRPCCallTimeout)
 	if err != nil {
 		log.Error("RPC Login failed: ", err)
 		r.ReAttemptLogin(err)
@@ -260,7 +262,7 @@ func (r *RPCStorageHandler) Login() {
 		return
 	}
 
-	ok, err := r.Client.Call("Login", r.UserKey)
+	ok, err := r.Client.CallTimeout("Login", r.UserKey, GlobalRPCCallTimeout)
 	if err != nil {
 		log.Error("RPC Login failed: ", err)
 		r.ReAttemptLogin(err)
@@ -294,7 +296,7 @@ func (r *RPCStorageHandler) GetKey(keyName string) (string, error) {
 	}
 
 	// Not cached
-	value, err := r.Client.Call("GetKey", r.fixKey(keyName))
+	value, err := r.Client.CallTimeout("GetKey", r.fixKey(keyName), GlobalRPCCallTimeout)
 
 	if err != nil {
 		if r.IsAccessError(err) {
@@ -324,7 +326,7 @@ func (r *RPCStorageHandler) GetRawKey(keyName string) (string, error) {
 
 func (r *RPCStorageHandler) GetExp(keyName string) (int64, error) {
 	log.Debug("GetExp called")
-	value, err := r.Client.Call("GetExp", r.fixKey(keyName))
+	value, err := r.Client.CallTimeout("GetExp", r.fixKey(keyName), GlobalRPCCallTimeout)
 
 	if err != nil {
 		if r.IsAccessError(err) {
@@ -348,7 +350,7 @@ func (r *RPCStorageHandler) SetKey(keyName string, sessionState string, timeout 
 		Timeout:      timeout,
 	}
 
-	_, err := r.Client.Call("SetKey", ibd)
+	_, err := r.Client.CallTimeout("SetKey", ibd, GlobalRPCCallTimeout)
 
 	if r.IsAccessError(err) {
 		r.Login()
@@ -368,7 +370,7 @@ func (r *RPCStorageHandler) SetRawKey(keyName string, sessionState string, timeo
 // Decrement will decrement a key in redis
 func (r *RPCStorageHandler) Decrement(keyName string) {
 	log.Warning("Decrement called")
-	_, err := r.Client.Call("Decrement", keyName)
+	_, err := r.Client.CallTimeout("Decrement", keyName, GlobalRPCCallTimeout)
 	if r.IsAccessError(err) {
 		r.Login()
 		r.Decrement(keyName)
@@ -384,7 +386,7 @@ func (r *RPCStorageHandler) IncrememntWithExpire(keyName string, expire int64) i
 		Expire:  expire,
 	}
 
-	val, err := r.Client.Call("IncrememntWithExpire", ibd)
+	val, err := r.Client.CallTimeout("IncrememntWithExpire", ibd, GlobalRPCCallTimeout)
 
 	if r.IsAccessError(err) {
 		r.Login()
@@ -414,7 +416,7 @@ func (r *RPCStorageHandler) GetKeysAndValuesWithFilter(filter string) map[string
 	searchStr := r.KeyPrefix + r.hashKey(filter) + "*"
 	log.Debug("[STORE] Getting list by: ", searchStr)
 
-	kvPair, err := r.Client.Call("GetKeysAndValuesWithFilter", searchStr)
+	kvPair, err := r.Client.CallTimeout("GetKeysAndValuesWithFilter", searchStr, GlobalRPCCallTimeout)
 
 	if r.IsAccessError(err) {
 		r.Login()
@@ -434,7 +436,7 @@ func (r *RPCStorageHandler) GetKeysAndValuesWithFilter(filter string) map[string
 func (r *RPCStorageHandler) GetKeysAndValues() map[string]string {
 
 	searchStr := r.KeyPrefix + "*"
-	kvPair, err := r.Client.Call("GetKeysAndValues", searchStr)
+	kvPair, err := r.Client.CallTimeout("GetKeysAndValues", searchStr, GlobalRPCCallTimeout)
 
 	if r.IsAccessError(err) {
 		r.Login()
@@ -455,7 +457,7 @@ func (r *RPCStorageHandler) DeleteKey(keyName string) bool {
 
 	log.Debug("DEL Key was: ", keyName)
 	log.Debug("DEL Key became: ", r.fixKey(keyName))
-	ok, err := r.Client.Call("DeleteKey", r.fixKey(keyName))
+	ok, err := r.Client.CallTimeout("DeleteKey", r.fixKey(keyName), GlobalRPCCallTimeout)
 
 	if r.IsAccessError(err) {
 		r.Login()
@@ -470,7 +472,7 @@ func (r *RPCStorageHandler) DeleteKey(keyName string) bool {
 
 // DeleteKey will remove a key from the database without prefixing, assumes user knows what they are doing
 func (r *RPCStorageHandler) DeleteRawKey(keyName string) bool {
-	ok, err := r.Client.Call("DeleteRawKey", keyName)
+	ok, err := r.Client.CallTimeout("DeleteRawKey", keyName, GlobalRPCCallTimeout)
 
 	if r.IsAccessError(err) {
 		r.Login()
@@ -492,7 +494,7 @@ func (r *RPCStorageHandler) DeleteKeys(keys []string) bool {
 		}
 
 		log.Debug("Deleting: ", asInterface)
-		ok, err := r.Client.Call("DeleteKeys", asInterface)
+		ok, err := r.Client.CallTimeout("DeleteKeys", asInterface, GlobalRPCCallTimeout)
 
 		if r.IsAccessError(err) {
 			r.Login()
@@ -542,7 +544,7 @@ func (r *RPCStorageHandler) AppendToSet(keyName string, value string) {
 		Value:   value,
 	}
 
-	_, err := r.Client.Call("AppendToSet", ibd)
+	_, err := r.Client.CallTimeout("AppendToSet", ibd, GlobalRPCCallTimeout)
 	if r.IsAccessError(err) {
 		r.Login()
 		r.AppendToSet(keyName, value)
@@ -560,7 +562,7 @@ func (r *RPCStorageHandler) SetRollingWindow(keyName string, per int64, val stri
 		Expire:  -1,
 	}
 
-	intVal, err := r.Client.Call("SetRollingWindow", ibd)
+	intVal, err := r.Client.CallTimeout("SetRollingWindow", ibd, GlobalRPCCallTimeout)
 	if r.IsAccessError(err) {
 		r.Login()
 		return r.SetRollingWindow(keyName, per, val)
@@ -587,7 +589,7 @@ func (r *RPCStorageHandler) SetRollingWindowPipeline(keyName string, per int64, 
 		Expire:  -1,
 	}
 
-	intVal, err := r.Client.Call("SetRollingWindow", ibd)
+	intVal, err := r.Client.CallTimeout("SetRollingWindow", ibd, GlobalRPCCallTimeout)
 	if r.IsAccessError(err) {
 		r.Login()
 		return r.SetRollingWindow(keyName, per, val)
@@ -635,7 +637,7 @@ func (r *RPCStorageHandler) GetApiDefinitions(orgId string, tags []string) strin
 		Tags:  tags,
 	}
 
-	defString, err := r.Client.Call("GetApiDefinitions", dr)
+	defString, err := r.Client.CallTimeout("GetApiDefinitions", dr, GlobalRPCCallTimeout)
 
 	if err != nil {
 		if r.IsAccessError(err) {
@@ -656,7 +658,7 @@ func (r *RPCStorageHandler) GetApiDefinitions(orgId string, tags []string) strin
 
 // GetPolicies will pull Policies from the RPC server
 func (r *RPCStorageHandler) GetPolicies(orgId string) string {
-	defString, err := r.Client.Call("GetPolicies", orgId)
+	defString, err := r.Client.CallTimeout("GetPolicies", orgId, GlobalRPCCallTimeout)
 	if err != nil {
 		if r.IsAccessError(err) {
 			r.Login()
@@ -717,14 +719,14 @@ func (r *RPCStorageHandler) CheckForKeyspaceChanges(orgId string) {
 	var err error
 
 	if config.SlaveOptions.GroupID == "" {
-		keys, err = r.Client.Call("GetKeySpaceUpdate", orgId)
+		keys, err = r.Client.CallTimeout("GetKeySpaceUpdate", orgId, GlobalRPCCallTimeout)
 	} else {
 
 		grpReq := GroupKeySpaceRequest{
 			OrgID:   orgId,
 			GroupID: config.SlaveOptions.GroupID,
 		}
-		keys, err = r.Client.Call("GetGroupKeySpaceUpdate", grpReq)
+		keys, err = r.Client.CallTimeout("GetGroupKeySpaceUpdate", grpReq, GlobalRPCCallTimeout)
 	}
 
 	if err != nil {